+++ changelog.md
@@ -6,6 +6,7 @@
 * `New` Omit parameter hints when the argument name matches
 * `FIX` Fix a typo in `no-unknown` diagnostic message
 * `FIX` Autodoc generation so it does not include documentation for builtin Lua language features
+* `FIX` Fix incorrect version in `view document`
 
 ## 3.16.1
 `2025-12-8`